Update transform functions to use float types and enhance output handling in LTO

+++ cpp/benchmarks/binaryop/fragments/mul_f32.cu
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-extern "C" __device__ int transform(int32_t* out, int32_t a, int32_t b) { *out = a * b; return 0; }
+extern "C" __device__ int transform(float* out, float a, float b) { *out = a * b; return 0; }
